new simulation pull series starting from different time points of a simulation at x=0. nm. The reason for this exercise is that the orientation gets locked the closer we approach the boundary and we want to test wether we can sample more orientations of the protein or whether the orientation is given by the protein surface interaction itself.

+++ md/06_pulls/launch_sim.py
@@ -5,6 +5,8 @@ import numpy as np
 this_folder = os.path.dirname(os.path.realpath(__file__))
 sys.path.append(this_folder + "/../scripts/")
 import gmx_tools.xvg
+import logging
+import time
 
 def do(cmd, comment=""):
    if len(comment) > 0:
@@ -13,12 +15,12 @@ def do(cmd, comment=""):
    os.system(cmd)
 
 if __name__ == "__main__":
-   parser = argparse.ArgumentParser(description='Launch simulations')
-   parser.add_argument('-i', '--input-folder', type=str, help="input folder where the pulling was run from", nargs='?', required=True)
+   parser = argparse.ArgumentParser(description='Launch simulations: pulls from a range of time frames of a single previous simulation')
+   parser.add_argument('-i', '--input-folder', type=str, help="input folder where the previous simulation was run from", nargs='?', required=True)
    parser.add_argument('-o', '--output-folder', type=str, help="output folder", nargs='?', required=True)
    parser.add_argument('-n', '--number-of-steps', type=int, help="number of sampling steps", required=True)
-   parser.add_argument('-s', '--min-distance-nm', type=float, help="minimum distance in unit nanometer", default=0.0)
-   parser.add_argument('-d', '--max-distance-nm', type=float, help="maximum distance in unit nanometer", required=True)
+   parser.add_argument('-s', '--min-time-ps', type=float, help="start sampling interval at time in unit ps")
+   parser.add_argument('-d', '--max-time-ps', type=float, help="end sampling interval at time in unit ps")
    parser.add_argument('-N', '--node', type=str, help="node")
    parser.add_argument('-G', '--gpus', type=int, help="number of GPUs", default=4)
 
@@ -26,12 +28,20 @@ if __name__ == "__main__":
       parser.print_help()
       sys.exit(0)
    args = parser.parse_args()
-
+   
    output_folder = os.path.abspath(args.output_folder)
    input_folder = os.path.abspath(args.input_folder)
 
-   do("rm -rf %s" % output_folder)
+   if os.path.exists(output_folder):
+      logging.warning('Output folder %s exists and is being overwritten.' % output_folder)
+      time.sleep(2)
+      do("rm -rf %s" % output_folder)
    do("mkdir %s" % output_folder)
+   # Logging call and all input argument of this script
+   with open("%s/launch_sim.log" % (output_folder), "w") as f:
+      for s in sys.argv:
+         f.write(s + " ")
+      f.write("\n")
    do("mkdir %s/input" % output_folder)
    
    xvg_filename = input_folder + "/out/pullx.xvg"
@@ -45,30 +55,40 @@ if __name__ == "__main__":
    tpr_filename = output_folder + "/input/sim.tpr"
    top_filename = output_folder + "/input/sim.top"
       
-   [time, xcom, xset] = gmx_tools.xvg.read_xvg(xvg_filename, ncols=3)
-    
-   x_grid = np.linspace(args.min_distance_nm, args.max_distance_nm, args.number_of_steps)
-   t_grid = np.interp(x_grid, xcom, time)
+   [times, xcom, xset] = gmx_tools.xvg.read_xvg(xvg_filename, ncols=3)
+
+   if args.min_time_ps is None:
+      logging.warning('No argument for --min-time-ps provided. Using first time point: %f ps.' % times[0])
+      time.sleep(2)
+      min_time_ps = times[0]
+   else:
+      min_time_ps = args.min_time_ps
+      
+   if args.max_time_ps is None:
+      logging.warning('No argument for --max-time-ps provided. Using last time point: %f ps.' % times[-1])
+      time.sleep(2)
+      max_time_ps = times[-1]      
+   else:
+      max_time_ps = args.max_time_ps
+   
+   t_grid = np.linspace(min_time_ps, max_time_ps, args.number_of_steps)
 
    exs = ["00_init", 
           "01_sim_pre",
           "02_sim"]
    cps = exs + ["itp", "mdp", "scripts", "log_to_file"]
    
-   for x,t in zip(x_grid, t_grid):
-      output_folder_sub = output_folder + ("/snap_%2.2fnm" % x)
+   for t in t_grid:
+      output_folder_sub = output_folder + ("/pull_from_snap_%06ips" % (int(round(t))))
       i_log = 0
       logfile = "%s/run.log.%i" % (output_folder_sub, i_log)
       # Do necessary initialisations
       do("mkdir %s" % output_folder_sub)
       do("mkdir %s/out" % output_folder_sub)
-      do("mkdir %s/input" % output_folder_sub)
-      do("cp %s %s/input/snap.top" % (top_filename, output_folder_sub))
+      do("cp %s %s/out/sim.top" % (top_filename, output_folder_sub))
       do("ln -sf run.log.%i %s/run.log" % (i_log, output_folder_sub), comment="Writing symlink to new log file")
       # Set up environment
       do("cp env %s/" % output_folder_sub)
-      with open("%s/env" % output_folder_sub, "a") as f:
-         f.write("GMXCPX_D0=\"%f\"\n" % x)
       # Write bassh script file to be submitted
       lines = []
       lines += ["#!/bin/bash\n",
@@ -77,9 +97,9 @@ if __name__ == "__main__":
                 "#SBATCH --nodes=1\n",
                 "#SBATCH --ntasks-per-node=1\n",
                 #"#SBATCH --partition=c\n",
-                "#SBATCH --partition=a\n",
+                #"#SBATCH --partition=a\n",
                 "#SBATCH --exclusive\n",
-                "#SBATCH --job-name=gmx%02.1f\n" % (round(x,1)),
+                "#SBATCH --job-name=gmx%03ius\n" % (int(round(t/1000))),
                 "#SBATCH --output=%s\n" % logfile,
                 "#SBATCH --workdir=%s\n" % output_folder_sub]
       if args.gpus > 0:
@@ -89,7 +109,7 @@ if __name__ == "__main__":
       lines.append("source ./env\n")
       for cp in cps:
          lines.append("cp -r -L %s/%s %s/\n" % (this_folder, cp, output_folder_sub))
-      lines.append("echo 0 | gmx trjconv -dump %f -f %s -s %s -o %s/out/snap.gro\n" % (t, trr_filename, tpr_filename, output_folder_sub))
+      lines.append("echo 0 | gmx trjconv -dump %f -f %s -s %s -o %s/out/sim.gro\n" % (t, trr_filename, tpr_filename, output_folder_sub))
       lines += [("echo \"./%s\"; ./%s\n") % (ex,ex) for ex in exs]
 
       slurm_file = "%s/run.sh" % output_folder_sub
